organized project structure and started using existing vocab
