+++ Datasets/WMT2014/WMT2014.swift
@@ -7,17 +7,11 @@
 
 import Foundation
 import TensorFlow
-
+import ModelSupport
 
 let BOS_WORD = "<s>"
 let EOS_WORD = "</s>"
 let BLANK_WORD = "<blank>"
-
-public func subsequentMask(size: Int) -> Tensor<Int32> {
-    let attentionShape = [1, size, size]
-    return Tensor<Int32>(ones: TensorShape(attentionShape))
-        .bandPart(subdiagonalCount: 0, superdiagonalCount: -1)
-}
 
 public struct WMT2014EnDe {
     public let directoryURL: URL
@@ -30,9 +24,9 @@ public struct WMT2014EnDe {
     public typealias ExampleIterator = IndexingIterator<[Example]>
     public typealias RepeatExampleIterator = ShuffleIterator<RepeatIterator<ExampleIterator>>
     public typealias TrainDataIterator = PrefetchIterator<
-        GroupedIterator<MapIterator<RepeatExampleIterator, TextBatch>>
+        GroupedIterator<MapIterator<RepeatExampleIterator, TranslationBatch>>
     >
-    public typealias DevDataIterator = GroupedIterator<MapIterator<ExampleIterator, TextBatch>>
+    public typealias DevDataIterator = GroupedIterator<MapIterator<ExampleIterator, TranslationBatch>>
 //    private typealias TestDataIterator = DevDataIterator
 
     public var trainDataIterator: TrainDataIterator
@@ -72,68 +66,8 @@ extension WMT2014EnDe {
 }
 
 extension WMT2014EnDe {
-    /// Tokenized text passage.
-    public struct TextBatch: KeyPathIterable {
-        /// IDs that correspond to the vocabulary used while tokenizing.
-        /// The shape of this tensor is `[batchSize, maxSequenceLength]`.
-        public var tokenIds: Tensor<Int32> // TODO: !!! Mutable in order to allow for batching.
-        // aka src
-        
-        public var targetTokenIds: Tensor<Int32>
-        // aka tgt
-        
-        /// IDs of the token types (e.g., sentence A and sentence B in BERT).
-        /// The shape of this tensor is `[batchSize, maxSequenceLength]`.
-        //    public var tokenTypeIds: Tensor<Int32> // TODO: !!! Mutable in order to allow for batching.
-        
-        /// Mask over the sequence of tokens specifying which ones are "real" as opposed to "padding".
-        /// The shape of this tensor is `[batchSize, maxSequenceLength]`.
-        public var mask: Tensor<Float> // TODO: !!! Mutable in order to allow for batching.
-        
-        public var targetMask: Tensor<Float> // TODO: !!! Mutable in order to allow for batching.
-        
-        public var targetTruth: Tensor<Int32>
-        
-        public var tokenCount: Int32
-        
-        public init(source: Tensor<Int32>, target: Tensor<Int32>, sourcePadId: Int32, targetPadId: Int32) {
-            self.tokenIds = source
-            self.mask = Tensor<Float>(Tensor(zerosLike: source)
-                .replacing(with: Tensor(onesLike: source), where: source .!= Tensor.init(sourcePadId))
-                .expandingShape(at: 1))
-            
-            let rangeExceptLast = 0..<(target.shape[1] - 1)
-            self.targetTokenIds = target[0...,rangeExceptLast]
-            self.targetTruth = target[0..., 1...]
-            self.targetMask = TextBatch.makeStandardMask(target: self.targetTokenIds, pad: targetPadId)
-            self.tokenCount = Tensor(zerosLike: targetTruth)
-                .replacing(with: Tensor(onesLike: targetTruth), where: self.targetTruth .!= Tensor.init(targetPadId))
-                .sum().scalar!
-            
-        }
-        
-        public init(tokenIds: Tensor<Int32>, targetTokenIds: Tensor<Int32>, mask: Tensor<Float>, targetMask: Tensor<Float>, targetTruth: Tensor<Int32>, tokenCount: Int32) {
-            self.tokenIds = tokenIds
-            self.targetTokenIds = targetTokenIds
-            self.mask = mask
-            self.targetMask = targetMask
-            self.targetTruth = targetTruth
-            self.tokenCount = tokenCount
-        }
-        
-        static func makeStandardMask(target: Tensor<Int32>, pad: Int32) -> Tensor<Float> {
-            var targetMask = Tensor(zerosLike: target)
-                .replacing(with: Tensor(onesLike: target), where: target .!= Tensor.init(pad))
-                .expandingShape(at: -2)
-            targetMask *= subsequentMask(size: target.shape.last!)
-            return Tensor<Float>(targetMask)
-        }
-    }
-}
-
-extension WMT2014EnDe {
     public init(
-        mapExample: @escaping (Example) -> TextBatch,
+        mapExample: @escaping (Example) -> TranslationBatch,
         taskDirectoryURL: URL,
     maxSequenceLength: Int,
     batchSize: Int) throws {
@@ -171,8 +105,8 @@ extension WMT2014EnDe {
                  reduceFn: WMT2014EnDe.reduceDataBatches(_:))
     }
     
-    static func reduceDataBatches(_ batches: [TextBatch]) -> TextBatch {
-        return TextBatch(tokenIds: Tensor(batches.map{ $0.tokenIds.squeezingShape(at: 0) }), // this should be fine
+    static func reduceDataBatches(_ batches: [TranslationBatch]) -> TranslationBatch {
+        return TranslationBatch(tokenIds: Tensor(batches.map{ $0.tokenIds.squeezingShape(at: 0) }), // this should be fine
                          targetTokenIds: Tensor(batches.map{ $0.targetTokenIds.squeezingShape(at: 0) }),
                          mask: Tensor(batches.map{ $0.mask.squeezingShape(at: 0) }),
                          targetMask: Tensor(batches.map{ $0.targetMask.squeezingShape(at: 0) }),
